requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
flask>=3.0.0
requests>=2.31.0
watchdog>=4.0.0
python-dotenv>=1.0.0
